pyproject.toml
[project]
name = "office-todo-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "sqlalchemy",
    "alembic",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "streamlit",
    "httpx",
    "websockets",
    "psycopg2-binary>=2.9.11",
    "argon2-cffi>=25.1.0",
    "pytest>=9.0.2",
    "black>=26.3.1",
    "flake8>=7.3.0",
]
